+++ NEWS.md
@@ -2,8 +2,11 @@
 
 * `set_tag()` skips if the tag already exists.
 * `create_package()` sets a code of conduct and contributing guidelines.
-* Run `pkgdown::build_site()` during the [`check_pkg`](https://github.com/inbo/actions/) GitHub action.
-* Deploy the `pkgdown` website to a `gp-pages` branch when pushing to master during the [`check_pkg`](https://github.com/inbo/actions/) GitHub action.
+* `create_package()` sets `LICENSE.md`.
+* Run `pkgdown::build_site()` during the 
+  [`check_pkg`](https://github.com/inbo/actions/) GitHub action.
+* Deploy the `pkgdown` website to a `gp-pages` branch when pushing to master 
+  during the [`check_pkg`](https://github.com/inbo/actions/) GitHub action.
 
 # checklist 0.1.3